refactor the abstract factory by facade
