--- src/main/java/org/patterns/abstractFactory/abstractInterface/Message.java
@@ -1,5 +1,0 @@
-package org.patterns.abstractFactory.abstractInterface;
-
-public interface Message {
-    void send(String message);
-}
