+++ src/main/java/org/patterns/abstractFactory/facade/abstractProducts/Message.java
@@ -1,0 +1,5 @@
+package org.patterns.abstractFactory.facade.abstractProducts;
+
+public interface Message {
+    void send(String message);
+}
