+++ src/main/java/org/patterns/abstractFactory/facade/concreteProducts/internet/InternetCommunicationFactory.java
@@ -1,0 +1,24 @@
+package org.patterns.abstractFactory.facade.concreteProducts.internet;
+
+import org.patterns.abstractFactory.facade.factory.CommunicationFactory;
+import org.patterns.abstractFactory.facade.abstractProducts.Message;
+import org.patterns.abstractFactory.facade.abstractProducts.VideoCall;
+import org.patterns.abstractFactory.facade.abstractProducts.VoiceCall;
+
+public class InternetCommunicationFactory implements CommunicationFactory {
+
+    @Override
+    public Message createMessage() {
+        return new WhatsApp();
+    }
+
+    @Override
+    public VideoCall makeVideoCall() {
+        return new Zoom();
+    }
+
+    @Override
+    public VoiceCall makeVoiceCall() {
+        return new VoIP();
+    }
+}
