+++ src/main/java/org/patterns/abstractFactory/facade/enums/CommunicationTypes.java
@@ -1,4 +1,4 @@
-package org.patterns.abstractFactory.enums;
+package org.patterns.abstractFactory.facade.enums;
 
 public enum CommunicationTypes {
     MOBILE, INTERNET
