--- src/main/java/org/patterns/abstractFactory/factory/CommunicationFactory.java
@@ -1,15 +1,0 @@
-package org.patterns.abstractFactory.factory;
-
-import org.patterns.abstractFactory.abstractInterface.Message;
-import org.patterns.abstractFactory.abstractInterface.VideoCall;
-import org.patterns.abstractFactory.abstractInterface.VoiceCall;
-
-public interface CommunicationFactory {
-
-    Message createMessage();
-
-    VideoCall makeVideoCall();
-
-    VoiceCall makeVoiceCall();
-
-}
